+++ packages/components/organisms/f-offers/src/components/_tests/NoOffersFound.test.js
@@ -1,16 +1,48 @@
-import { mount } from '@vue/test-utils';
+import { createLocalVue, mount } from '@vue/test-utils';
 import MediaElement from '@justeat/f-media-element';
 import Button from '@justeat/f-button';
+import { VueI18n } from '@justeat/f-globalisation/src';
+import Vuex from 'vuex';
 import NoOffersFound from '../NoOffersFound.vue';
 import tenantConfigs from '../../tenants';
+import { VUEX_MODULE_NAMESPACE_OFFERS } from '../../store/types';
+
+const localVue = createLocalVue();
+
+localVue.use(VueI18n);
+localVue.use(Vuex);
 
 const mockLocale = 'en-GB';
+
+const i18n = {
+    locale: 'en-GB',
+    messages: {
+        'en-GB': tenantConfigs['en-GB'].messages
+    }
+};
 
 describe('NoOffersFound.vue', () => {
     let wrapper;
     beforeEach(() => {
         // Arrange
+        const getters = {
+            isAuthenticated: () => false,
+            friendlyName: () => null
+        };
+
+        const store = new Vuex.Store({
+            modules: {
+                [VUEX_MODULE_NAMESPACE_OFFERS]: {
+                    namespaced: true,
+                    getters
+                }
+            }
+        });
+
         wrapper = mount(NoOffersFound, {
+            store,
+            i18n,
+            localVue,
             mocks: {
                 $t: key => key.split('.').reduce((acc, current) => acc[current], tenantConfigs[mockLocale].messages)
             },